Move ava-specific handling to base & fix profiling
