+++ models/common/base.py
@@ -6,6 +6,7 @@ from pytorch_lightning.utilities.parsing import AttributeDict
 from ride.core import Configs, RideMixin
 from ride.utils.logging import getLogger
 from ride.utils.utils import name
+from datasets.ava import ava_loss, preprocess_ava_batch
 
 logger = getLogger("co3d")
 
@@ -100,6 +101,10 @@ class Co3dBase(RideMixin):
         if "frame" in self.hparams.co3d_forward_mode:
             self.module.call_mode = "forward_steps"  # default = "forward"
 
+        if self.hparams.dataset == "ava":
+            self.loss = ava_loss(self.loss)
+            self.hparams.enable_detection = True
+
         logger.info(f"Model receptive field: {self.module.receptive_field} frames")
         logger.info(f"Training loss: {name(self.loss)}")
 
@@ -124,8 +129,24 @@ class Co3dBase(RideMixin):
                 if hasattr(m, "stride_index"):
                     m.stride_index = 0
 
+            if self.hparams.enable_detection:
+                # Pass in bounding boxes to RoIHead for AVA dataset.
+                dummy_boxes = [torch.tensor([[11.5200, 25.0880, 176.0000, 250.6240]])]
+                self.module[-1].set_boxes(dummy_boxes)
+
+    def preprocess_batch(self, batch):
+        """Overloads method in ride.Lifecycle"""
+        if self.hparams.dataset == "ava":
+            batch = preprocess_ava_batch(batch)
+        return batch
+
     def forward(self, x):
         result = None
+
+        if self.hparams.enable_detection and not self.hparams.profile_model:
+            # Pass in bounding boxes to RoIHead for AVA dataset.
+            self.module[-1].set_boxes(x["boxes"])
+            x = x["images"]
 
         if "init" in self.hparams.co3d_forward_mode:
             self.module.clean_state()
