+++ models/coresnet/main.py
@@ -7,7 +7,7 @@ from ride.optimizers import SgdOneCycleOptimizer
 from ride.utils.logging import getLogger
 
 from datasets import ActionRecognitionDatasets
-from datasets.ava import AvaMetric, ava_loss, preprocess_ava_batch
+from datasets.ava import AvaMetric
 from models.common import Co3dBase
 from models.coresnet.modules.coresnet import CoResNet
 
@@ -109,10 +109,6 @@ class CoResNetRide(
         return c
 
     def __init__(self, hparams):
-        if self.hparams.dataset == "ava":
-            self.loss = ava_loss(self.loss)
-            self.hparams.enable_detection = True
-
         self.module = CoResNet(
             arch=self.hparams.resnet_architecture,
             dim_in=self.dim_in,
@@ -130,19 +126,6 @@ class CoResNetRide(
             align_detection=self.hparams.align_detection,
             temporal_fill=self.hparams.co3d_temporal_fill,
         )
-
-    def preprocess_batch(self, batch):
-        """Overloads method in ride.Lifecycle"""
-        if self.hparams.dataset == "ava":
-            batch = preprocess_ava_batch(batch)
-        return batch
-
-    def forward(self, x):
-        if self.hparams.enable_detection:
-            # Pass in bounding boxes to RoIHead.
-            self.module[-1].set_boxes(x["boxes"])
-            x = x["images"]
-        return Co3dBase.forward(self, x)
 
     def map_loaded_weights(self, finetune_from_weights, state_dict):
         # Map state_dict for "Slow" weights
